Initial Commit
While there's still a lot more that I would like to do, I'm simply making a backup of my work at this point in time. I've not even published this library to NPM & I don't wish to publish this package until I feel like this library is in a more mature state.